pyproject.toml
[project]
name = "whisper-asr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20250625",
    "yt-dlp>=2025.11.12",
]
